+++ pages/_app.tsx
@@ -1,18 +1,13 @@
 import React from 'react';
 import type { AppProps } from 'next/app';
 import { Hydrate, QueryClient, QueryClientProvider } from 'react-query';
-import Nav from '@/components/Nav';
-import Footer from '@/components/Footer';
+import Nav from '@/components/layout/Nav';
+import Footer from '@/components/layout/Footer';
 import { ReactQueryDevtools } from 'react-query/devtools';
 
 import '@/styles/global.css';
+import '@/styles/tailwind.style';
 
-// core styles shared by all of react-notion-x (required)
-import 'react-notion-x/src/styles.css';
-// used for code syntax highlighting (optional)
-import 'prismjs/themes/prism-tomorrow.css';
-// used for rendering equations (optional)
-import 'katex/dist/katex.min.css';
 import { RecoilRoot } from 'recoil';
 
 export default function CustomApp({ Component, pageProps }: AppProps) {
